input: nt36523_spi: don't export any sysfs nodes at all
These cause a kernel panic due to the new version of sec_input already exporting most of these.
It seems like we don't need these, and a look at the same driver found in the a54x kernel source shows that the sec_fn part of the driver isn't even built.

Change-Id: I3a76427e4f5cbb69dc2cc9acfa5a33577b2ce1c9
Signed-off-by: Nahuel Gómez <nahuelgomez329@gmail.com>
Signed-off-by: Majaahh <maja.blaszczak@proton.me>