sound: soc/abox: Do not perform a direct firmware load
When using a unified vendor that relies on bind mounts to select the correct
firmware per device, a race condition occurs where the firmware tries to load
before the mounts are present (and the firmware is available).

Instead of request_firmware_direct() and having the kernel read the firmware
on its own, we can use request_firmware() to tell ueventd to load the firmware
when it's available. This gives a much wider margin of error.

Signed-off-by: Flopster101 <nahuelgomez329@gmail.com>
Signed-off-by: Ksawlii <maja.blaszczak@proton.me>